Move "env" into "helpers" module
