+++ scripts/export-dos-participants.py
@@ -7,9 +7,8 @@ Usage:
 import sys
 sys.path.insert(0, '.')
 
-import csv
 from docopt import docopt
-from dmscripts.env import get_api_endpoint_from_stage
+from dmscripts.helpers.env import get_api_endpoint_from_stage
 from dmscripts.export_dos_suppliers import (
     FRAMEWORK_SLUG, find_services_by_lot, make_fields_from_content_questions, write_csv
 )
