+++ scripts/generate-buyer-email-list.py
@@ -10,7 +10,7 @@ sys.path.insert(0, '.')
 
 from docopt import docopt
 from dmapiclient import DataAPIClient
-from dmscripts.env import get_api_endpoint_from_stage
+from dmscripts.helpers.env import get_api_endpoint_from_stage
 from dmscripts.generate_buyer_email_list import list_buyers
