+++ scripts/oneoff/rename-dos-locations.py
@@ -12,7 +12,7 @@ from multiprocessing.pool import ThreadPool
 import itertools
 from docopt import docopt
 from dmapiclient import DataAPIClient
-from dmscripts.env import get_api_endpoint_from_stage
+from dmscripts.helpers.env import get_api_endpoint_from_stage
 
 
 def create_draft_getter(client):
