+++ scripts/oneoff/rename-west-england.py
@@ -10,10 +10,9 @@ sys.path.insert(0, '.')
 
 from multiprocessing.pool import ThreadPool
 import itertools
-import json
 from docopt import docopt
 from dmapiclient import DataAPIClient
-from dmscripts.env import get_api_endpoint_from_stage
+from dmscripts.helpers.env import get_api_endpoint_from_stage
 
 OLD_LOCATION = "West England"
 NEW_LOCATION = "South West England"
