+++ scripts/set-search-alias.py
@@ -6,10 +6,11 @@ Usage:
 """
 
 import sys
+
 from docopt import docopt
 
 sys.path.insert(0, '.')
-from dmscripts.env import get_api_endpoint_from_stage
+from dmscripts.helpers.env import get_api_endpoint_from_stage
 from dmapiclient import SearchAPIClient
